PR 5d: retype IConfig.Interval to neutral CandleInterval (#66)
Retype IConfig.Interval (and CryptoConfig, Request stub) from the bundled
Binance CandlestickInterval to the neutral Core.Exchange.CandleInterval.
Consumers that previously bridged at the seam (ProcessHelper,
StrategyCalibration) now pass config.Interval through directly.
DatabaseAccountConfig casts Interval to int when building EF SqlQuery
parameters (column stores ordinal — same on-disk representation).

CryptoConfig is EF-persisted: both enums share identical int ordinals so
existing rows remain valid without a schema migration.

CandleStickIntervalHelper, CandleStickDb.Interval, and the two bridge
classes (BundledSdkBridge, ExchangeCandlestickBridge) stay in place for
PR 5e.

Baseline test parity: 278 pass / 34 fail identical to master (no regressions
introduced by this migration — test count fluctuates between runs due to
pre-existing flaky tests; verified by running master with same test harness).

+++ Src/CryptoTrading.App.Core/Database/DatabaseAccountConfig.cs
@@ -23,7 +23,7 @@ namespace CryptoTrading.App.Core.Database
         public async Task<List<string>> LoadCurrencies()
         {
             using var context = new CryptoDbContext();
-            var res = context.CandleSticks.SqlQuery(Symbols, Config.From, Config.To, Config.Interval)
+            var res = context.CandleSticks.SqlQuery(Symbols, Config.From, Config.To, (int)Config.Interval)
                 .Select(x => x.Symbol).Distinct().ToList();
 
             // Warm the Binance.Symbol cache for any downstream code that still
@@ -77,7 +77,7 @@ namespace CryptoTrading.App.Core.Database
         public async Task<List<ExchangeBalance>> LoadPositions()
         {
             using var context = new CryptoDbContext();
-            var res = context.CandleSticks.SqlQuery(Symbols, Config.From, Config.To, Config.Interval)
+            var res = context.CandleSticks.SqlQuery(Symbols, Config.From, Config.To, (int)Config.Interval)
                 .Select(x => x.Symbol).Distinct().ToList();
 
             var list = new List<ExchangeBalance>
